+++ src/dram/Dram.cc
@@ -306,7 +306,8 @@ void Dram::dump(std::ostream &os) const
 {
 	os << "\n\n--------------------\n\n";
 	os << "Dumping DRAM system\n";
-	os << misc::fmt("%d Controllers\nController dump:\n", controllers.size());
+	os << misc::fmt("%d Controllers\nController dump:\n",
+			(int) controllers.size());
 
 	for (auto controller : controllers)
 	{
